Move stream safe wrapping responsibility to `ConfigurationCacheIO`
So the whole stream pipeline construction can be in a single place in
preparation for the introduction of compression.
